// client/src/interfaces/user.interfaces.ts
export interface User {
  id: number;
  username: string;
  password: string;
  password_confirm: string;
  email: string;
  gender: string;
  online_status: string;
  last_name: string;
  first_name: string;
  last_online_date: string;
  registration_date: string;
  photo: string;
  phone: string;
  birthday: string;
}